% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphanorm.R
\name{coef.alphanorm}
\alias{coef.alphanorm}
\title{Output the coefficients of "alphanorm" object}
\usage{
\method{coef}{alphanorm}(alphanorm.obj)
}
\arguments{
\item{alphanorm.obj}{a fitted "alphanorm" object}
}
\value{
coefficients of "alphanorm" object
}
\description{
Output the coefficients of "alphanorm" object
}
\seealso{
\code{\link{alphanorm}}
}
