% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphanorm.R
\name{cv.alphanorm}
\alias{cv.alphanorm}
\title{Cross-validation for alpha-norm}
\usage{
cv.alphanorm(x, y, lambda_Tune = exp(-10:10), q_Tune = c(0.1, 0.5, 0.9),
  intercept = TRUE, nfolds = 5, tol = 1e-07, T = 500)
}
\arguments{
\item{x}{design matrix}

\item{y}{response vector}

\item{lambda_Tune}{user-supplied lambda sequence}

\item{q_Tune}{user-supplied q sequence}

\item{intercept}{whether intercept should be in the model, default to be TRUE}

\item{nfolds}{number of folds , default to be 5}

\item{tol}{tolerence of convergence condition}

\item{T}{number of maximum iterations for each coefficient}
}
\value{
An object of S3 class "cv.alphanorm"
\item{lambda}{the values of lambda used in the fits}
\item{q}{the values of q used in the fits}
\item{cvm}{The mean cross-validation error, a matrix of length(q)*length(lambda) }
\item{lambda.min}{value of lambda that gives minimum cvm}
\item{q.min}{value of q that gives minimum cvm}
}
\description{
Does k-fold cross-validation for alpha-norm, and return the best lambda and q
}
\seealso{
\code{\link{alphanorm}}
}
