% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphanorm.R
\name{plot.alphanorm}
\alias{plot.alphanorm}
\title{plot coefficient for a "alphanorm"}
\usage{
\method{plot}{alphanorm}(alphanorm.obj, xvar = c("lambda"), legend = FALSE)
}
\arguments{
\item{alphanorm.obj}{fitted "alphanorm" model}

\item{xvar}{what is on the X-axis. "norm" plots against the $L_q$-norm of the coefficients, "lambda" against the log-lambda sequence}

\item{legend}{whether legend should be plotted}
}
\description{
Produce a coefficient profile plot of the coefficient paths for a fitted "alphanorm" object
}
\examples{
x=matrix(rnorm(100*20),100,20)
y=rnorm(100)
obj1=alphanorm(x,y)
plot(obj1)
plot(obj1,xvar="norm")
}
\seealso{
\code{\link{alphanorm}}
}
