% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphanorm.R
\name{predict.alphanorm}
\alias{predict.alphanorm}
\title{Predict method for alpha-norm fits}
\usage{
\method{predict}{alphanorm}(alphanorm.obj, newx = NULL)
}
\arguments{
\item{alphanorm.obj}{a fitted alpha-norm model, returned by alphanorm()}

\item{newx}{matrix of new values of x, if NULL, use the x in alphanorm.obj}
}
\value{
matrix of fitted values from alpha-norm model
}
\description{
Similar to other predict methods, this function predicts fitted values from a fitted alphanorm model
}
\seealso{
alphanorm,and cv.alphanorm methods
}
